requirements.txt
configargparse
transformers
openai
torch
pillow
accelerate
bitsandbytes